chore: organized app files
