--- client/src/views/app/Blog.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import HomeLayout from "../../layout/HomeLayout.jsx";
-import Blog from "../../pages/Blog/Blog.jsx";
-
-function BlogPage() {
-  return (
-    <HomeLayout>
-     
-      <Blog />
-    
-
-    </HomeLayout>
-  );
-}
-
-export default BlogPage;
