+++ client/src/views/app/Blog/index.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import HomeLayout from "../../../layout/HomeLayout.jsx";
+import Blog from "../../../pages/Blog/Blog.jsx";
+
+function BlogPage() {
+  return (
+    <HomeLayout>
+     
+      <Blog />
+    
+
+    </HomeLayout>
+  );
+}
+
+export default BlogPage;
