+++ client/src/views/app/ContributorPage/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import HomeLayout from "../../layout/HomeLayout.jsx";
-import Contact from "../../pages/Contact/Contact.jsx";
+import HomeLayout from "../../../layout/HomeLayout.jsx";
+import Contact from "../../../pages/Contact/Contact.jsx";
 
 function ContributorPage() {
   return (
