+++ client/src/views/app/Customizer/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import EditorLayout from '../../../layout/EditorLayout'
+
+function Customizer() {
+  return (
+    <EditorLayout>Customizer</EditorLayout>
+  )
+}
+
+export default Customizer
