+++ client/src/views/app/Faq/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import HomeLayout from "../../layout/HomeLayout.jsx";
-import FaqPage from "../../pages/Faq/Faq.jsx";
+import HomeLayout from "../../../layout/HomeLayout.jsx";
+import FaqPage from "../../../pages/Faq/Faq.jsx";
 
 function Faq() {
   return (
