--- client/src/views/app/Guide.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-import HomeLayout from "../../layout/HomeLayout.jsx";
-import GuideBook from "../../pages/Guidebook/guidebook.jsx";
-function Guide() {
-  return (
-    <HomeLayout>
-     
-      <GuideBook />
-    
-
-    </HomeLayout>
-  );
-}
-
-export default Guide;
