+++ client/src/views/app/Guide/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import HomeLayout from "../../../layout/HomeLayout.jsx";
+import GuideBook from "../../../pages/Guidebook/guidebook.jsx";
+function Guide() {
+  return (
+    <HomeLayout>
+      <GuideBook />
+    </HomeLayout>
+  );
+}
+
+export default Guide;
